Display current location and pins on map.
Loads JSON-stored locations and attempts to display them on the map. Unfortunately I am still unable to squash a persistent and very strange bug. The getCurrentLocation API method does not run on any of my browsers (webkit, Gecko). It used to work until yesterday. As a result of this bug, curLatDeg and curLngDeg are "undefined", which is why the map does not get drawn. This is *not* a problem of variable scope, as far as I can tell. Might be an API quirk? Help would be appreciated.