All well now—Have a nice trip to Oman
All works well. Event listener added so that map is certainly in place before attempting to place pins. Pins are added normally. Confusion was largely due to the sample locations being… close to Oman. I was looking for the in the Helsinki area and it did not occur me to zoom the map out and look for them at some other part of the globe. Imbecility, but now it's over. Avanti!