Added infoWindows (aka "bubbles") to pins.
Clicking on a pin now opens a "bubble" that will contain event information. The text is editable inline. We still need to determine whether this is sufficient, or whether a dialog to add/edit event information is still the way to go. My hunch is that editable bubbles should be enough.